+++ OrrRock/OrrRock/ViewControllers/HomeViewController/HomeViewController.swift
@@ -56,26 +56,7 @@ final class HomeViewController : UIViewController {
             }
         }
     }
-    
-    //인디게이터 사용을 위한 선언
-    private let indicator : NVActivityIndicatorView = {
-        let view = NVActivityIndicatorView(frame: CGRect(x: 0, y: 0, width: 150, height: 150),
-                                           type: .lineSpinFadeLoader,
-                                           color: .lightGray,
-                                           padding: 50)
-        view.backgroundColor = UIColor.black.withAlphaComponent(0.8)
-        view.layoutIfNeeded()
-        view.layer.cornerRadius = 15
-        view.layer.masksToBounds = true
-        return view
-    }()
-    
-    //인디게이터가 돌 때 다른 동작을 못하게 하기위한 뷰
-    private lazy var blockTouchView: UIView = {
-        let view = UIView()
-        return view
-    }()
-    
+
     private lazy var headerView: UIView = {
         let view = UIView()
         
@@ -282,24 +263,6 @@ final class HomeViewController : UIViewController {
             $0.leading.equalTo(view.safeAreaLayoutGuide.snp.leading)
             $0.trailing.equalTo(view.safeAreaLayoutGuide.snp.trailing)
         }
-
-
-        // 인디게이터 위치 추가
-        self.view.addSubview(indicator)
-        indicator.snp.makeConstraints {
-            $0.center.equalTo(self.view)
-        }
-    }
-    
-    //이친구는 반복적으로 추가되고 삭제 되어야해서 따로 만들었습니다.
-    func addBlockTouchView() {
-        self.view.addSubview(blockTouchView)
-        blockTouchView.snp.makeConstraints {
-            $0.top.equalTo(view.snp.top)
-            $0.trailing.equalTo(view.snp.trailing)
-            $0.leading.equalTo(view.snp.leading)
-            $0.bottom.equalTo(view.snp.bottom)
-        }
     }
     
     private func setUICollectionViewDelegate() {
@@ -310,73 +273,12 @@ final class HomeViewController : UIViewController {
     @objc func switchViewStyle() {
         self.isCardView.toggle()
     }
-}
 
-extension HomeViewController {
-    
     @objc func videoButtonPressed(sender: UIButton){
         let nextVC = SetDateViewController()
         self.navigationController?.pushViewController(nextVC, animated: true)
     }
-    
-//    @objc func videoButtonPressed(sender: UIButton!) {
-//        var configuration = PHPickerConfiguration()
-//        configuration.selectionLimit = 0
-//        //인디게이터 도는거 보고 싶으면 아랫줄을 주석 처리해주세요.
-//        configuration.preferredAssetRepresentationMode = .current
-//        configuration.filter = .any(of: [.videos])
-//        let picker = PHPickerViewController(configuration: configuration)
-//        picker.delegate = self
-//        self.present(picker, animated: true, completion: nil)
-//    }
-    
-    //터치를 제한하는 뷰를 추가하고 인디게이터를 실행 시킵니다.
-    func startIndicator() {
-        addBlockTouchView()
-        indicator.startAnimating()
-    }
-    
-    func stopIndicator() {
-        self.indicator.stopAnimating()
-        blockTouchView.removeFromSuperview()
-    }
-}
 
-extension HomeViewController: PHPickerViewControllerDelegate {
-    
-    func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
-        picker.dismiss(animated: true)
-        var videoUrlArray: [URL] = []
-        //인디케이트를 소환합니다.
-        startIndicator()
-        
-        //사용자가 영상을 선택 하지 않은 상태일 때
-        if results.count == 0 {
-            //인디게이터 종료
-            stopIndicator()
-        }
-        
-        //선택된 영상에서 URL을 뽑아내는 로직입니다.
-        for i in 0..<results.count {
-            results[i].itemProvider.loadFileRepresentation(forTypeIdentifier: UTType.movie.identifier) { url, err in
-                if url == nil {
-                    NSLog("Orr_HomeViewController_Err1:\(String(describing: err))\n")
-                } else {
-                    videoUrlArray.append(url!)
-                }
-
-                if results.count - 1 == i {
-                    DispatchQueue.main.sync {
-                        //인디케이터 종료
-                        self.stopIndicator()
-                        let nextVC = UpoadTestNextViewController()
-                        nextVC.viewUrlArray = videoUrlArray
-                        self.navigationController?.pushViewController(nextVC, animated: true)
-                    }
-                }
-            }
-        }
-    }
 }
 
 // QuickAction을 통한 정렬 및 필터링 시 함수를 아래에 구현
